requirements.txt
flask
flask-cors
gunicorn
python-doctr
opencv-python
numpy
pillow
pandas
openpyxl
